web-payment-service

+++ modules/currency-service/src/main/java/ua/boring/project/currencyservice/service/impl/CurrencyServiceImpl.java
@@ -3,6 +3,9 @@ package ua.boring.project.currencyservice.service.impl;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ua.boring.project.currencyservice.data.dto.CurrencyDto;
+import ua.boring.project.currencyservice.data.entity.CurrencyList;
+import ua.boring.project.currencyservice.data.repository.CurrencyListRepository;
 import ua.boring.project.currencyservice.utils.CurrencyCalculator;
 import ua.boring.project.currencyservice.utils.CurrencyParser;
 import ua.boring.project.currencyservice.service.CurrencyService;
@@ -11,16 +14,21 @@ import ua.boring.project.currencyservice.data.repository.CurrencyRepository;
 
 import java.io.IOException;
 import java.math.BigDecimal;
+import java.util.Map;
 
 @Slf4j
 @Service
 public class CurrencyServiceImpl implements CurrencyService {
 
     private final CurrencyRepository currencyRepository;
+//    private final CurrencyListRepository currencyListRepository;
 
     @Autowired
-    public CurrencyServiceImpl(CurrencyRepository currencyRepository) {
+    public CurrencyServiceImpl(CurrencyRepository currencyRepository
+//                               CurrencyListRepository currencyListRepository
+    ) {
         this.currencyRepository = currencyRepository;
+//        this.currencyListRepository = currencyListRepository;
     }
 
     @Override
@@ -44,10 +52,24 @@ public class CurrencyServiceImpl implements CurrencyService {
         }
 
         try {
-            Currency currency = CurrencyParser.getCurrencies();
+            Currency currencyData = CurrencyParser.getCurrencies();
+//
+//            Currency currency = new Currency();
+//            currency.setBase(currencyData.getBase());
+//            currency.setDate(currencyData.getDate());
+//
+//            for (Map.Entry<String, BigDecimal> entry : currencyData.getRates().entrySet()) {
+//
+//                CurrencyList currencyList = new CurrencyList();
+//                currencyList.setDate(currencyData.getDate());
+//                currencyList.setCurrencyKey(entry.getKey());
+//                currencyList.setCurrencyValue(entry.getValue());
+//
+//                currencyListRepository.save(currencyList);
+//            }
 
-            currencyRepository.save(currency);
-        } catch (IOException e) {
+            currencyRepository.save(currencyData);
+        } catch (Exception e) {
             log.debug("Exception while saving new currencies: ", e);
 
             return -1;
@@ -64,8 +86,11 @@ public class CurrencyServiceImpl implements CurrencyService {
     @Override
     public Integer removeLast() {
         try {
-            Currency currency = currencyRepository.findTopByOrderByIdDesc();
-            currencyRepository.delete(currency);
+            Currency currency = getLast();
+
+            if (currency != null) {
+                currencyRepository.delete(currency);
+            }
 
             return 0;
         } catch (Exception ex) {
